+++ demo/scripts/index-static.js
@@ -11,7 +11,9 @@ export default ({
   baseUrl,
 }) => (
   statinamicStatic({
-    pkg,
+    metadata: { 
+      pkg,
+    },
     urls,
     pagesData,
     dest,