+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 
 - Fixed: `peerDependencies` have been adjusted to minimal requirements
+- Added: `MetadataProvider` to send the data you want to the React context
+in a `metadata` attribute.
 
 # 0.1.0 - 2015-10-21